refactor(cron): split lifecycle discovery and task execution helpers
